feat(P4T3): T6.4 Strategy Comparison Tool and Risk Dashboard DB Integration (#82)

+++ strategies/alpha_baseline/backtest.py
@@ -185,7 +185,7 @@ class PortfolioBacktest:
 
             daily_returns.append(portfolio_return)
 
-        return pd.Series(daily_returns, index=dates[: len(daily_returns)])
+        return pd.Series(daily_returns, index=dates[: len(daily_returns)], dtype=float)
 
     def _compute_metrics(self) -> dict[str, float]:
         """
